Projeto - 86acj3fhz -> Classes de movimentações criadas, entidades Vendedores e Usuários com os códigos refatorados e terminados.

+++ Projeto-LojaDeMateriais/src/main/core/java/br/com/rpinfo/lorenzo/domain/repositories/produtos/ProdutoDao.java
@@ -12,7 +12,7 @@ public interface ProdutoDao{
 
     boolean update (Produtos produto) throws Exception;
 
-    Produtos getProduto (Long id) throws Exception;
+    Produtos getProduto (Integer id) throws Exception;
 
     List<Produtos> getListProdutos () throws Exception;
 }
